refactor: 객체의 키를 한글로 사용하는 것을 switch로 대체함
